+++ src/module/setupModules.ts
@@ -6,6 +6,7 @@ const modules = {
   "about-time": "0.0", 
   "anonymous": "0.0.0",
   "combat-utility-belt": "1.3.8",
+  "condition-lab-triggler": "1.4",
   "dae": "10.0.30",
   "ddb-game-log": "0.0.0",
   "df-templates": "1.0.0",
@@ -40,7 +41,6 @@ export let setupModules = () => {
       if (game.modules.get(name)?.active) {
         //@ts-ignore game.module.version
         error(`midi-qol requires ${name} to be of version ${modules[name]} or later, but it is version ${game.modules.get(name)?.version}`);
-        if (name === "dfreds-convenient-effects") ui.notifications?.warn(`You must upgrade convenient effects t0 4.0.2 or later`, {permanent: true});
       } else console.warn(`midi-qol | optional module ${name} not active - some features disabled`)
     }
   }
@@ -58,8 +58,7 @@ export function dice3dEnabled() {
 
 export function checkModules() {
   if (game.user?.isGM && !installedModules.get("socketlib")) {
-    //@ts-ignore expected one argument but got 2
-    ui.notifications.error("midi-qol.NoSocketLib", {permanent: true, localize: true});
+    ui.notifications?.error("midi-qol.NoSocketLib", {permanent: true, localize: true});
   }
   //@ts-ignore
   const midiVersion = game.modules.get("midi-qol").version;
@@ -69,23 +68,8 @@ export function checkModules() {
 }
 
 export function checkCubInstalled() {
-  return;
-  if (game.user?.isGM && configSettings.concentrationAutomation && !installedModules.get("combat-utility-belt")) {
-    let d = new Dialog({
-      // localize this text
-      title: i18n("midi-qol.confirm"),
-      content: i18n("midi-qol.NoCubInstalled"), 
-      buttons: {
-          one: {
-              icon: '<i class="fas fa-check"></i>',
-              label: "OK",
-              callback: ()=>{
-                configSettings.concentrationAutomation = false;
-              }
-          }
-      },
-      default: "one"
-    })
-    d.render(true);
+  //@ts-expect-error game.version
+  if (game.user?.isGM && installedModules.get("combat-utility-belt") && isNewerVersion(game.version, "11.0")) {
+    ui.notifications?.warn("midi-qol.cubNotSupported", {permanent: true, localize: true})
   }
 }